# openapi-pisp-initiator.yaml
openapi: 3.0.3
info:
  title: "PISP Initiator API (Open Finance v5)"
  description: >
    API exposta pela Iniciadora (PISP) e intermediada pelo Sensedia APIM.
    Expõe os endpoints nativos da iniciadora — iniciação de pagamento com
    redirect (PISP v5) e jornadas JSR (cadastro de dispositivo + pagamento
    sem redirect) — além dos paths formais Open Finance de autenticação,
    contas e consentimentos.

    Jornada JSR:
      - POST /enrollments inicia o cadastro do dispositivo (ITP).
      - GET /callback recebe o redirect da detentora e conclui o cadastro
        (gera credential_id e vincula a conta).
      - POST /payments/jsr inicia um pagamento PIX sem redirect usando o
        dispositivo vinculado (exige device REGISTERED; caso contrário
        retorna need_enrollment com o login_url).
  version: "1.1.0"
servers:
  - url: "https://api-assets.sensedia.com/v1"
    description: "APIM Sensedia"
tags:
  - name: "JSR"
    description: "Jornada JSR (ITP enrollment + pagamento sem redirect)"
  - name: "PISP"
    description: "Iniciação de Pagamento"
  - name: "Auth"
    description: "Autenticação"
  - name: "Accounts"
    description: "Contas"
paths:
  # ---------------------------------------------------------------------------
  # Endpoints nativos da Iniciadora — Jornada JSR
  # ---------------------------------------------------------------------------

  /enrollments:
    post:
      tags: [JSR]
      summary: "Inicia o cadastro de um dispositivo (JSR/ITP) na detentora"
      operationId: "createEnrollment"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentRequest"
      responses:
        "201":
          description: "Cadastro iniciado; abra login_url para autenticar na detentora"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-api-broken: false
      x-interceptors: {}

  /callback:
    get:
      tags: [JSR]
      summary: "Recebe o redirect de volta da detentora e completa o fluxo em curso"
      description: >
        Completa o cadastro do dispositivo (kind=enrollment) ou o pagamento
        com redirect. Para enrollment, retorna credential_id e account_id.
      operationId: "handleCallback"
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Fluxo concluído (resultado varia conforme o tipo)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallbackResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
      x-api-broken: false
      x-interceptors: {}

  /payments/jsr:
    post:
      tags: [JSR]
      summary: "Inicia um pagamento PIX sem redirect (JSR) usando o dispositivo vinculado"
      description: >
        Exige um dispositivo REGISTERED. Se não houver dispositivo, retorna
        HTTP 400 com need_enrollment=true e o login_url para cadastro.
      operationId: "createJsrPayment"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsrPaymentRequest"
      responses:
        "201":
          description: "Pagamento iniciado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsrPaymentResponse"
        "400":
          description: "Nenhum dispositivo vinculado — cadastro necessário"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeedEnrollmentResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-api-broken: false
      x-interceptors: {}

  # ---------------------------------------------------------------------------
  # Endpoints nativos da Iniciadora — Pagamento com redirect
  # ---------------------------------------------------------------------------

  /payments:
    post:
      tags: [PISP]
      summary: "Inicia um pagamento PIX com redirect na detentora"
      operationId: "createPayment"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitiationRequest"
      responses:
        "201":
          description: "Pagamento iniciado; abra login_url para autenticar"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStartedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-api-broken: false
      x-interceptors: {}

  /payments/{identifier}:
    get:
      tags: [PISP]
      summary: "Consulta o status de um pagamento por payment_id ou consent_id"
      operationId: "getPaymentDetail"
      parameters:
        - name: identifier
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Detalhes do pagamento"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
      x-api-broken: false
      x-interceptors: {}

  # ---------------------------------------------------------------------------
  # Paths formais Open Finance (superfície exposta; roteamento configurado
  # diretamente no portal Sensedia — sem x-destination no spec)
  # ---------------------------------------------------------------------------

  /open-banking/auth/authorize:
    post:
      tags: [Auth]
      summary: "Proxy: inicia o fluxo OAuth da detentora"
      operationId: "proxyAuthAuthorize"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                redirect_uri:
                  type: string
              required: [redirect_uri]
      responses:
        "200":
          description: "Auth request criado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                  login_url:
                    type: string
      x-api-broken: false
      x-interceptors: {}

  /open-banking/auth/token:
    post:
      tags: [Auth]
      summary: "Proxy: troca code por access_token"
      operationId: "proxyAuthToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
              required: [code]
      responses:
        "200":
          description: "Token emitido"
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
      x-api-broken: false
      x-interceptors: {}

  /open-banking/accounts:
    get:
      tags: [Accounts]
      summary: "Proxy: lista contas do usuário autenticado"
      operationId: "proxyListAccounts"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Lista de contas"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
      x-api-broken: false
      x-interceptors: {}

  /open-banking/pisp/payments/v5/consents:
    post:
      tags: [PISP]
      summary: "Cria um consentimento de pagamento PIX (PISP v5)"
      operationId: "createPispV5Consent"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PispV5ConsentRequest"
      responses:
        "201":
          description: "Consentimento criado"
          headers:
            x-pisp-consent-id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      x-api-broken: false
      x-interceptors: {}

  /open-banking/pisp/payments/v5/pix/payments:
    post:
      tags: [PISP]
      summary: "Inicia o pagamento PIX após autorização"
      operationId: "initiatePispV5Payment"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                consentId:
                  type: string
                endToEndIds:
                  type: array
                  items:
                    type: string
              required: [consentId]
      responses:
        "201":
          description: "Pagamento iniciado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitiatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      x-api-broken: false
      x-interceptors: {}

  /open-banking/pisp/payments/v5/pix/payments/{paymentId}:
    get:
      tags: [PISP]
      summary: "Consulta o status de um pagamento PIX"
      operationId: "getPispV5PaymentStatus"
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Status do pagamento"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      x-api-broken: false
      x-interceptors: {}

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # ---- JSR: Enrollment ----
    EnrollmentRequest:
      type: object
      properties:
        username:
          type: string
          description: "Nome do titular a vincular (JSR/ITP)"
        account_number:
          type: string
          description: "Conta de débito do titular (opcional)"
      required: [username]

    EnrollmentCreatedResponse:
      type: object
      properties:
        enrollment_id:
          type: string
        request_id:
          type: string
        login_url:
          type: string
        message:
          type: string

    CallbackResponse:
      type: object
      description: >
        Resultado do callback. Para enrollment: enrollment_id, credential_id,
        account_id, status. Para pagamento: payment_id, consent_id, status.
      properties:
        enrollment_id:
          type: string
        credential_id:
          type: string
        account_id:
          type: string
        payment_id:
          type: string
        consent_id:
          type: string
        status:
          type: string

    # ---- JSR: Pagamento ----
    JsrPaymentRequest:
      type: object
      properties:
        amount:
          type: string
          description: "Valor em BRL, ex: '100.00'"
        currency:
          type: string
          default: BRL
        creditor_name:
          type: string
        creditor_cpf_cnpj:
          type: string
        creditor_key:
          type: object
          description: "{type, value} da chave PIX do credor"
          properties:
            type:
              type: string
              enum: [CPF, CNPJ, EMAIL, PHONE, EVP]
            value:
              type: string
          required: [type, value]
      required: [amount, creditor_name, creditor_cpf_cnpj, creditor_key]

    JsrPaymentResponse:
      type: object
      properties:
        payment_id:
          type: string
        consent_id:
          type: string
        status:
          type: string

    NeedEnrollmentResponse:
      type: object
      properties:
        need_enrollment:
          type: boolean
          example: true
        login_url:
          type: string
        message:
          type: string

    # ---- Pagamento com redirect (native) ----
    PaymentInitiationRequest:
      type: object
      properties:
        amount:
          type: string
          description: "Valor em BRL, ex: '100.00'"
        currency:
          type: string
          default: BRL
        creditor_name:
          type: string
        creditor_cpf_cnpj:
          type: string
        creditor_key:
          type: object
          description: "{type, value} da chave PIX do credor"
          properties:
            type:
              type: string
              enum: [CPF, CNPJ, EMAIL, PHONE, EVP]
            value:
              type: string
          required: [type, value]
        account_id:
          type: string
          nullable: true
          description: "Override da conta de débito"
      required: [amount, creditor_name, creditor_cpf_cnpj, creditor_key]

    PaymentStartedResponse:
      type: object
      properties:
        request_id:
          type: string
        login_url:
          type: string
        message:
          type: string

    PaymentDetailResponse:
      type: object
      properties:
        consent_id:
          type: string
        request_id:
          type: string
        status:
          type: string
        payment_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    # ---- PISP v5 (formal) ----
    PispV5ConsentRequest:
      type: object
      properties:
        redirect_uri:
          type: string
          nullable: true
        authorisation_server:
          type: object
          properties:
            authorisation_server_id:
              type: string
            organisation_id:
              type: string
        creditor:
          type: object
          properties:
            person_type:
              type: string
              enum: [PESSOA_NATURAL, PESSOA_JURIDICA]
            cpf_cnpj:
              type: string
            name:
              type: string
        payment:
          type: object
          properties:
            type:
              type: string
              default: PIX
            purpose:
              type: string
              enum: [IMMEDIATE]
            date:
              type: string
              format: date
            currency:
              type: string
              default: BRL
            amount:
              type: string
            details:
              type: object
              properties:
                local_instrument:
                  type: string
                  default: DICT
                proxy:
                  type: string
                creditor_account:
                  type: object
                  properties:
                    ispb:
                      type: string
                    issuer:
                      type: string
                    number:
                      type: string
                    account_type:
                      type: string
        debtor_account:
          type: object
          properties:
            ispb:
              type: string
            issuer:
              type: string
            number:
              type: string
            account_type:
              type: string
        remittance_information:
          type: string
      required: [creditor, payment, debtor_account]

    ConsentCreatedResponse:
      type: object
      properties:
        consent_id:
          type: string
        redirect_uri:
          type: string
        request_id:
          type: string

    PaymentInitiatedResponse:
      type: object
      properties:
        payment_id:
          type: string
        consent_id:
          type: string
        end_to_end_id:
          type: string
        status:
          type: string
        amount:
          type: string
        balance:
          type: string

    PaymentStatusResponse:
      type: object
      properties:
        consent_id:
          type: string
        status:
          type: string
        payment_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

  responses:
    BadRequest:
      description: "Requisição inválida"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
    Unauthorized:
      description: "Não autenticado"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
    NotFound:
      description: "Não encontrado"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
    InternalServerError:
      description: "Erro interno"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string